Allow scrolling of Table in ScrolledComposite
See Bug 385714 - [iOs] Touch-Scrolling a Table not possible within a
ScrolledComposite.
In contrast to the RAP-master, this does not allow
switching from Table-to-ScrolledComposite scrolling in one
touch-session. This is due to the "webkitOverflowScrolling = "touch""
style. After reaching the end of the Table, the user has to touch it
again to scroll the outer element.