Bug 383397 -Supporting unreliable networks by adding a request timeout handling
https://bugs.eclipse.org/bugs/show_bug.cgi?id=383397

After manually aborting a request it's not possible to read any
properties of the request object anymore. Every attempt will lead to a
failure in Request#_handleFailed() when reading the responseText. See
also 
http://stackoverflow.com/questions/7287706/ie-9-javascript-error-c00c023f.

However, connection aborts should be handled by
_handleConnectionError. Therefore the function isConnectionError
has been extended to return true in case of -1 which is set because
reading the state failed too.