Bug 383397 - Supporting unreliable networks by adding a request timeout handling
https://bugs.eclipse.org/bugs/show_bug.cgi?id=383397

org.eclipse.swt.Request is a singleton, therefore autoRetryCounter must
not be kept as instance variable without resetting it after a successful
request. Resetting the counter was missing which made the
messagebox appear too often. Since I'm not 100% sure about the order of
execution and whether there are concurrent requests or not
(requestCounter is confusing me), I bound the counter to the actual
request object instead of just resetting it in _handleCompleted.