Bug 383397 - Supporting unreliable networks by adding a request timeout handling
https://bugs.eclipse.org/bugs/show_bug.cgi?id=383397

- Sometimes, on Webkit browsers, a "ParseException - Unexpected end of
input" occured. This happened when aborting a connection while reading
the response. The readyState was 4, httpState was 200 but response
empty. Fixed by using a timeout event which always does a retry without
checking for the status first.
- On slow networks it happened that the response never reached the
client because it was too large and took longer to transfer than the
timeout was. Increased the timeout from 2000 to 9000.