Bug 425574 - ServerPush sometimes stops working
Possibility added to set timeouts:
1. Abort timeout on client side for ajax calls, when reached request
will be aborted
2. Wait timeout on server side for server pushes, when reached server
push request will be released and reinitiated on the client side.

Timeouts are not enabled by default. To enable use the following code on
the server side:
ServerPushManager.getInstance().setRequestTimeout(REQUEST_TIMEOUT);
ServerPushManager.getInstance().setRequestCheckInterval(REQUEST_TIMEOUT);

And the following js on the client side:
 // patch ServerPush_createRequest()
  rwt.qx.Class.__initializeClass( rwt.client.ServerPush );
  var proto = rwt.client.ServerPush.prototype;
  var orgFunc = proto._createRequest;
  proto._createRequest = function() {
    var request = orgFunc.apply( this, arguments );
    request.setTimeout(10000);
    return request;
  };